requirements.txt
"""
Requirements file for Email Management AI Agent.
Lists all required Python packages.
"""

# Core dependencies
flask==2.3.3
werkzeug==2.3.7
jinja2==3.1.2

# Google API dependencies
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0

# Email processing
nltk==3.8.1
python-dateutil==2.8.2

# Testing
pytest==7.4.0
unittest-mock==1.3.0
